Refactor gameplay lifecycle into deterministic stages
